#153 #157 #161 Fix crash when recreating holder activity
