+++ cardstackview/src/main/java/com/yuyakaido/android/cardstackview/CardStackView.java
@@ -6,53 +6,12 @@ import android.support.v7.widget.RecyclerView;
 import android.util.AttributeSet;
 import android.view.MotionEvent;
 
+import com.yuyakaido.android.cardstackview.internal.CardStackDataObserver;
 import com.yuyakaido.android.cardstackview.internal.CardStackSnapHelper;
 
 public class CardStackView extends RecyclerView {
 
-    private final AdapterDataObserver observer = new AdapterDataObserver() {
-        private CardStackLayoutManager getCardStackLayoutManager() {
-            LayoutManager manager = getLayoutManager();
-            if (manager instanceof CardStackLayoutManager) {
-                return (CardStackLayoutManager) manager;
-            }
-            throw new IllegalStateException("CardStackView must be set CardStackLayoutManager.");
-        }
-        @Override
-        public void onChanged() {
-            CardStackLayoutManager manager = getCardStackLayoutManager();
-            manager.setTopPosition(0);
-        }
-
-        @Override
-        public void onItemRangeChanged(int positionStart, int itemCount) {
-            // Do nothing
-        }
-
-        @Override
-        public void onItemRangeChanged(int positionStart, int itemCount, @Nullable Object payload) {
-            // Do nothing
-        }
-
-        @Override
-        public void onItemRangeInserted(int positionStart, int itemCount) {
-            // Do nothing
-        }
-
-        @Override
-        public void onItemRangeRemoved(int positionStart, int itemCount) {
-            CardStackLayoutManager manager = getCardStackLayoutManager();
-            if (positionStart == 0) {
-                manager.setTopPosition(0);
-            }
-            manager.removeAllViews();
-        }
-
-        @Override
-        public void onItemRangeMoved(int fromPosition, int toPosition, int itemCount) {
-            // Do nothing
-        }
-    };
+    private final CardStackDataObserver observer = new CardStackDataObserver(this);
 
     public CardStackView(Context context) {
         this(context, null);
@@ -90,6 +49,7 @@ public class CardStackView extends RecyclerView {
         RecyclerView.Adapter adapter = getAdapter();
         if (adapter != null) {
             adapter.registerAdapterDataObserver(observer);
+            observer.isRegistered = true;
         }
     }
 
@@ -98,7 +58,10 @@ public class CardStackView extends RecyclerView {
         super.onDetachedFromWindow();
         RecyclerView.Adapter adapter = getAdapter();
         if (adapter != null) {
-            adapter.unregisterAdapterDataObserver(observer);
+            if (observer.isRegistered) {
+                adapter.unregisterAdapterDataObserver(observer);
+                observer.isRegistered = false;
+            }
         }
     }
